define behavior for division by zero
previously, division by zero placed the strings "Infinity", "-Infinity",
or "NaN" into the first number property of the text display. This gave
the user a way to enter non-token strings to the text display object.
Now, an error message is displayed and the text display object is reset
upon a NaN or infinite result